rename(UserDomain): rename folder name to camelCase
