+++ snapnow-domain/src/main/java/kr/flab/snapnow/domain/user/enums/setting/SettingCategory.java
@@ -1,14 +1,16 @@
 package kr.flab.snapnow.domain.user.enums.setting;
 
 import java.util.EnumSet;
-import kr.flab.snapnow.domain.user.enums.setting.setting_option.NotificationSettingOption;
-import kr.flab.snapnow.domain.user.enums.setting.setting_option.PrivacySettingOption;
-import kr.flab.snapnow.domain.user.enums.setting.setting_option.SettingOption;
-import kr.flab.snapnow.domain.user.enums.setting.setting_type.NotificationSetting;
-import kr.flab.snapnow.domain.user.enums.setting.setting_type.PrivacySetting;
-import kr.flab.snapnow.domain.user.enums.setting.setting_type.SettingType;
+
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import kr.flab.snapnow.domain.user.enums.setting.settingOption.NotificationSettingOption;
+import kr.flab.snapnow.domain.user.enums.setting.settingOption.PrivacySettingOption;
+import kr.flab.snapnow.domain.user.enums.setting.settingOption.SettingOption;
+import kr.flab.snapnow.domain.user.enums.setting.settingType.NotificationSettingType;
+import kr.flab.snapnow.domain.user.enums.setting.settingType.PrivacySettingType;
+import kr.flab.snapnow.domain.user.enums.setting.settingType.SettingType;
+import kr.flab.snapnow.domain.user.exception.InvalidSettingException;
 
 @Getter
 @RequiredArgsConstructor
@@ -16,25 +18,16 @@ public enum SettingCategory {
 
     NOTIFICATION(
         "Notification",
-        EnumSet.allOf(NotificationSetting.class),
+        EnumSet.allOf(NotificationSettingType.class),
         EnumSet.allOf(NotificationSettingOption.class)),
     PRIVACY(
         "Privacy",
-        EnumSet.allOf(PrivacySetting.class),
+        EnumSet.allOf(PrivacySettingType.class),
         EnumSet.allOf(PrivacySettingOption.class));
 
     private final String value;
     private final EnumSet<? extends SettingType> settingType;
     private final EnumSet<? extends SettingOption> settingOption;
-
-    public static boolean isValidSetting(
-        Enum<? extends SettingType> settingType,
-        Enum<? extends SettingOption> settingOption) {
-        return (NOTIFICATION.getSettingType().contains((SettingType) settingType)
-                && NOTIFICATION.getSettingOption().contains((SettingOption) settingOption))
-                || (PRIVACY.getSettingType().contains((SettingType) settingType)
-                && PRIVACY.getSettingOption().contains((SettingOption) settingOption));
-    }
 
     public static boolean isValidSetting(SettingType settingType, SettingOption settingOption) {
         return NOTIFICATION.getSettingType().contains(settingType)
@@ -48,6 +41,31 @@ public enum SettingCategory {
         SettingType settingType,
         SettingOption settingOption) {
         return settingCategory.getSettingType().contains(settingType)
-            && settingCategory.getSettingOption().contains(settingOption);
+                && settingCategory.getSettingOption().contains(settingOption);
+    }
+
+    public static SettingCategory of(SettingType settingType) {
+        for (SettingCategory category : SettingCategory.values()) {
+            if (category.getSettingType().contains(settingType)) {
+                return category;
+            }
+        }
+        throw new InvalidSettingException();
+    }
+
+    public static SettingCategory of(SettingOption settingOption) {
+        for (SettingCategory category : SettingCategory.values()) {
+            if (category.getSettingOption().contains(settingOption)) {
+                return category;
+            }
+        }
+        throw new InvalidSettingException();
+    }
+
+    public static SettingCategory of(SettingType settingType, SettingOption settingOption) {
+        if (isValidSetting(settingType, settingOption)) {
+            return of(settingType);
+        }
+        throw new InvalidSettingException();
     }
 }
