+++ snapnow-domain/src/main/java/kr/flab/snapnow/domain/user/enums/setting/settingOption/NotificationSettingOption.java
@@ -1,4 +1,4 @@
-package kr.flab.snapnow.domain.user.enums.setting.setting_option;
+package kr.flab.snapnow.domain.user.enums.setting.settingOption;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -7,10 +7,11 @@ import kr.flab.snapnow.domain.user.enums.setting.SettingCategory;
 
 @Getter
 @RequiredArgsConstructor
-public enum PrivacySettingOption implements SettingOption {
+public enum NotificationSettingOption implements SettingOption {
 
-    PRIVATE("Private"),
-    PUBLIC("Public");
+    ALL("All"),
+    FROM_FOLLOWER("From Follower"),
+    OFF("Off");
 
     private final String value;
 
@@ -21,6 +22,6 @@ public enum PrivacySettingOption implements SettingOption {
 
     @Override
     public SettingCategory getCategory() {
-        return SettingCategory.PRIVACY;
+        return SettingCategory.NOTIFICATION;
     }
 }
