+++ snapnow-domain/src/main/java/kr/flab/snapnow/domain/user/enums/setting/settingType/SettingType.java
@@ -1,14 +1,13 @@
-package kr.flab.snapnow.domain.user.enums.setting.type;
+package kr.flab.snapnow.domain.user.enums.setting.settingType;
 
 import java.util.EnumSet;
 
 import kr.flab.snapnow.domain.user.enums.setting.SettingCategory;
-import kr.flab.snapnow.domain.user.enums.setting.option.SettingOption;
+import kr.flab.snapnow.domain.user.enums.setting.settingOption.SettingOption;
 
 public interface SettingType {
 
     String getCode();
-
     String getValue();
     EnumSet<? extends SettingOption> getOptions();
     SettingCategory getCategory();
