+++ snapnow-domain/src/main/java/kr/flab/snapnow/domain/user/model/userSetting/NotificationSetting.java
@@ -1,10 +1,10 @@
-package kr.flab.snapnow.domain.user.model.user_setting;
+package kr.flab.snapnow.domain.user.model.userSetting;
 
 import java.util.HashMap;
 
 import kr.flab.snapnow.domain.user.enums.setting.SettingCategory;
-import kr.flab.snapnow.domain.user.enums.setting.setting_option.NotificationSettingOption;
-import kr.flab.snapnow.domain.user.enums.setting.setting_type.NotificationSettingType;
+import kr.flab.snapnow.domain.user.enums.setting.settingOption.NotificationSettingOption;
+import kr.flab.snapnow.domain.user.enums.setting.settingType.NotificationSettingType;
 import kr.flab.snapnow.domain.user.exception.InvalidSettingException;
 
 public class NotificationSetting extends UserSetting<NotificationSettingType, NotificationSettingOption> {
