+++ snapnow-domain/src/main/java/kr/flab/snapnow/domain/user/model/userSetting/PrivacySetting.java
@@ -1,10 +1,10 @@
-package kr.flab.snapnow.domain.user.model.user_setting;
+package kr.flab.snapnow.domain.user.model.userSetting;
 
 import java.util.HashMap;
 
 import kr.flab.snapnow.domain.user.enums.setting.SettingCategory;
-import kr.flab.snapnow.domain.user.enums.setting.setting_option.PrivacySettingOption;
-import kr.flab.snapnow.domain.user.enums.setting.setting_type.PrivacySettingType;
+import kr.flab.snapnow.domain.user.enums.setting.settingOption.PrivacySettingOption;
+import kr.flab.snapnow.domain.user.enums.setting.settingType.PrivacySettingType;
 import kr.flab.snapnow.domain.user.exception.InvalidSettingException;
 
 public class PrivacySetting extends UserSetting<PrivacySettingType, PrivacySettingOption> {
