+++ snapnow-domain/src/main/java/kr/flab/snapnow/domain/user/model/userSetting/UserSetting.java
@@ -1,4 +1,4 @@
-package kr.flab.snapnow.domain.user.model.user_setting;
+package kr.flab.snapnow.domain.user.model.userSetting;
 
 import java.util.Map;
 
@@ -7,8 +7,8 @@ import lombok.experimental.SuperBuilder;
 import lombok.NoArgsConstructor;
 
 import kr.flab.snapnow.domain.user.enums.setting.SettingCategory;
-import kr.flab.snapnow.domain.user.enums.setting.setting_option.SettingOption;
-import kr.flab.snapnow.domain.user.enums.setting.setting_type.SettingType;
+import kr.flab.snapnow.domain.user.enums.setting.settingOption.SettingOption;
+import kr.flab.snapnow.domain.user.enums.setting.settingType.SettingType;
 
 @Getter
 @SuperBuilder
